requirements.txt
beautifulsoup4==4.13.4
fastapi==0.115.12
fastapi_cache==0.1.0
markdownify==1.1.0
openai==1.75.0
pydantic==2.11.3
python-dotenv==1.1.0
redis==4.6.0
Requests==2.32.3
uvicorn==0.34.2
